requirements.txt
pandas==2.2.2
numpy==1.24.4
prophet==1.1.5
pycaret==3.2.0
shap==0.44.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
vertica-python==1.3.0
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==1.7.6
